refactor: update functions used by script and remove prompts folder
